+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Anthropic: Support for [Structured Output](https://inspect.aisi.org.uk/structured.html) for Sonnet 4.5 and Opus 4.1.
+
 ## 0.3.146 (15 November 2025)
 
 - Added `cache` configuration to `GenerateConfig` (formerly was only available as a parameter to `generate()`).